requirements.txt
asgiref==3.6.0
Django==3.2.16
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pylama==8.4.1
pytz==2022.7.1
snowballstemmer==2.2.0
sqlparse==0.4.3
